requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
billiard==4.2.4
cachetools==6.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
fastapi==0.128.0
google-api-core==2.28.1
google-auth==2.45.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
packaging==25.0
prompt-toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
redis==7.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
